Branch for main development + reorganized files.
An actual (working) development version should be always in this
  branch.
For subtasks, new branches will be created. For now all the old
  files are here, although this is actually a subtask. Later,
  we will remove all the old code from this branch and move it to
  some other branch.
The master branch should be used for working or stable versions.

Added .gitignore file.
Created basic directory structure for the project and changed
  Makefile accordingly.
Added QtCreator project files (CuteDNS.*).