Changed rehashing so it always finishes + bugfix.
In case the stash is not big enough, it will get resized.
Only if it cannot be resized (allocation failure), the rehashing
  fails. (And no other rehashing is possible!!)
In future, provide some way to notify the administrator to reload
  the server, or provide a way to try to finish the rehash when
  some memory is available again.

Bugfix in us_hash() - the coeficient index was not counted well.
Removed 'inline' from dynamic array getters. (Should look into the
  inline stuff in future!)
Corrected some errors in debug output.