Recommitted RRSet adjusting.
This code was disabled long ago in a temporary commit dealing with
something totally different (f97bde66). Reenabled it, but I'm not
sure this won't break something else.

Also replaced the closest encloser being saved to the RDATA dname
by the wildcard child. It should prevent possible problems with
query processing. Although it is worth checking out if they are not
accustomed to this (closest encloser in the dname.node field).

refs #1419 @3h