Simplified TSIG recognition and fixed handling.
TSIG is now detected while parsing the RR sections. If it is
there, it is stored in the packet. This also allows for checking
whether TSIG is not in a forbidden place, or whether there is
only one TSIG, etc.

Also removed all TSIG recognition code from other parts of the
server. Not needed anymore.

refs #1522 @1h