Fixed problem with wrong handling of failed transfers.
This is probably a misconception. When transfer failed at some point,
the appropriate handling function would either send buffered response
and/or error message with specified rcode. But then the handling message
would free the response structure and caller of the function would attempt
to create and send ANOTHER error message based on the return value.
This would obviously fail hard, as the response structure is freed at this point.

refs #2176