Removed gentree from nodes for a simple array.
Reason is - we rarely get more than a few different RR
types in a node (usually <4). So it makes not much sense
to build an expensive AVL tree and allocate other stuff.
Worst case is O(n) (but probably less) which is OK for this
scenario. If we bothered to order the array on insertion,
we could get O(logN) with binary search then - no extra memory
costs.