Implemented RRL collsision checking and slow-start.
When a collision occurs (bucket is same, but addresses differ),
bucket enters a slow-start mode.
This means it is given less tokens for two seconds in a row,
then leaves the slow-start mode.
Reason for this is penalize collisions, but still allow
legitimate clients to connect.
In usual mode of operations collisions on the same bucket shouldn't
happen in a regular fashion, therefore not triggering slow-start
mode often.