Corrected privileges dropping in some cases.
The PID file needed to be created before with knotc to indicate
starting instance. This is better than knotd creating it, but
it could result in either dropping privileges before starting knotd
or creating PID file in an inaccessible directory.
As of now, knotd drops privileges and creates empty PID file before
loading zones.

Change-Id: Ifb1507b2a922217d4d4608accb1f71dd7be5c0ee