Simplified and cleaned up label compression.
The idea was to use simplified tries, but it didn't work so well.

Simplified and cleaned up label compression.

The idea was to use simplified tries, but it didn't work so well.
New approach is based on the fact, that only a small part of unique
suffixes is reused in the packet (mostly qname and few related names),
whilst rest of the names are mentioned scarcely.
As a tradeoff, we will compres those most common names.

Table is scanned in a linear fashion.
Name suffix is aligned (end-to-end) to a current table pointer.
Aligned labels are then compared and best coverage is reported.
Matches that can't exceed current best coverage are ignored.
Not/partially covered names are inserted (only index to wire).
Covering matches are promoted to a lower index, this is because:
- frequent matches are inspected earlier, performance increase
- last 1/4 of the least-used names may be replaced randomly with new names

Didn't change the currently used API to keep the patch nice
and sweet, but there's a lot of legacy code to be removed.

Change-Id: Id86bc7e3e56012182eebf148aad3ddc3b2649879