Backup & flush journal when zone file is update manually.
This prevents inconsistency when using DDNS and you update your zone
as well before data in journal is flushed.
To choice is to invalidate data in the journal. If one choses to
revert to previous zone, he has to revert changes in the zone file
and overwrite current journal with the backed up one.
Note that the journal has to have newer mtime.