Asynchronous connect for transfers, fixed self pacing.
It was possible for the previous implementation to sometimes overshoot
the hard limit by a few connections, this yielded very undesirable results
as the master starts dropping connections and forcing slave to reschedule
transfers.

Asynchronous connect is also faster as other pending transfers may be processed
during the connection phase.

refs #65