Fixed OPT RR/NSID malformed on reload and unsafe alteration.
Fixed a bug where NSID couldn't be switched off on reload, causing
malformed packets and/or extra bytes.
Since it's protected by RCU, it needs to be created on each server
reload and the old one should be freed after the RCU is synchronized.
Also removed rest of obsolete code and unified EDNS constants.