Fixed saving merged changeset.
Saving now done in store-apply-commit sequence. This ensures that
changes are not saved to journal if something went wrong.

There is still one problem about this - we apply only the DNSSEC
changeset and even if it's wrong, the changeset from zone diff
should be saved to the journal. Should fix this later.

refs #4