Descriptive log messages in the zone loading.
I still don't like it much. I'd like to have a separate function
that would always print current zone file, line and record so
the logs are predictable. That would require more refactoring
so the scanner is visible from all the functions and I'd like to
leave it as a part of a bigger change.