Fixed wrong change in RRSets counting in node.
Each RRSet that is stored in node, even if it's only a placeholder,
must be counted because the array of RRSets is indexed by this
count.

Changed condition in NSEC chain generation - the NSEC should be
dropped also in case where there are NSEC and some placeholder
RRSets in the node.

refs #210