query_module: configurable query processing modules
Current API provides a "query plan" for each zone, which is
basically a set of steps it needs to takes a query to be resolved.
The plan is divided into several stages that reflect answer section
processing. Now this query plan can be altered via configurable modules,
first module is a 'synth_record' that makes synthetic forward/reverse
records if the query isn't satisfied from the zone contents.