xfr_handler: packet parsing code unified, axfr fallback bugfix
The problem with previous code was that the packet stripped the TSIG
section when it parsed the first IXFR packet from wire. If the code
failed though, the wire was parsed again but this time the TSIG was
already stripped on the wire so the result was garbled data.
I've thrown in some minor cleanup as well.