RDATA lowercase: Converting in packet parsing + zone load.
Converting to lowercase when saving to zone proved insufficient,
because case-insensitive comparison of RDATA would be needed in some
scenarios - e.g. in IXFR (SOA replacement) and DDNS processing.

Moved the conversion to packet parsing - all dnames in RDATA
are converted to lowercase when parsed from wire. RDATA parsed
from zone file are converted there (in scanner_process()). Later
the conversion may be moved to a lower level of parsing (and with
some option to turn it on/off).

QNAME is unaffected by this change.