libdnssec/key: bad logic leaves key with dangling pointer
when setting key from rdata, a copy of binary data struct is made.
that struct is then resized to accomodate for new key.
if successful, binary data points to freed memory
when pubkey parsing fails, the resized copy is discarded
and key binary data left with dangling pointer
freeing key causes double-free and using it use-after-free