server: avoid duplicate NSEC for Wildcard No Data answer
The proof for Wildcard No Data consists of two NSEC records: The first
one matches the wildcard record and shows that the QTYPE doesn't exist.
The second one shows there is no better match.

If there are no names below the wildcard in the zone, these two NSEC
records will be the same.