docs: manually set copyright year in doc/conf.py
To make builds bit-by-bit reproducible, the source code should not
change what is built based on the current time.

Currently, the documentation sets its build year by querying the
system clock, which means that the produced documentation changes when
you build it in a different year.

This approach costs one manual commit per year of development, but
always builds the same thing, and claims no additional copyright years
if no changes have been made to the source.

It provides a nice warning so that (hopefully) someone building the
package which needs the manual year update should notice and can make
the update themselves if the source code is actually active (as i
expect it to be in 2020, but who knows what will happen in 2035?)

Compare with the use of SOURCE_DATE_EPOCH:
https://gitlab.labs.nic.cz/knot/knot-dns/merge_requests/951