geoip: add support for weighted records
Weighted records are a useful mechanism for situations where uneven
load-balancing is required, such as a canary environment or a gradual
roll-out.

Instead of using deterministic properties like location or source IP,
allow to attach a statistical weight to a certain record. The ratio of
a record's weight to the sum of the weights of all records for a domain
determines the record's likelyhood to be the result of a random pick.