geoip: trigger CNAME chain resolution
Currently, when the geoip module returns a CNAME, knot will not resolve
it any further, even if it is within its authority. This forces clients
or recursors to issue an additional request to resolve the CNAME.

This commit enables the module to take advantage of knot's chain
resolution by updating the query data with the returned CNAME and
returning KNOT_IN_STATE_FOLLOW instead of KNOT_IN_STATE_HIT, which will
let knot to continue with the processing as it would for regular CNAME
results.