tests-extra: random DDNS updates -- don't count apex NS RRset removals as changes
Complete apex NS RRSet removals (ANY NS removals) aren't processed by the server
(see RFC 2136, Section 7.13). If such a removal (or removals) is the only change
in the update, then it's a no-change DDNS update to the zone, which usually breaks
the test. Therefore, generate such removals, but don't count them as changes.