backup: create a backup label and check for its existence
The label file identifies the backup an prevents another backup to be made
to an already existing backup. Existence of a label file is necessary for a restore,
which helps to prevent a frequent user error and to avoid a number of failures
which would be triggered by it.

In addition to that, every label file contains a brief description
of the backup and its parameters. A list of zones that the backup contains
isn't included so far.