backup: get the backup format version to be used by restore
Iff there isn't a label file and the force flag is set, use the original backup
format for restore.

Otherwise report a backup without a label file as malformed.

If there is an empty label file, the label file doesn't start with a label file
header, the label file doesn't include a backup format description, or the backup
format code isn't known, consider it a malformed backup and don't restore from it,
even if the force flag has been set.