mod-dnstap: Add 'responses-with-queries' option
This commit adds a 'responses-with-queries' option to the dnstap module.
When enabled, it also adds the original query message to AUTH_RESPONSE
type dnstap message payloads in addition to the response message.

The dnstap protobuf definition has separate 'query_message' and
'response_message' fields for dnstap message objects, and for
AUTH_RESPONSE messages the 'query_message' field can optionally be
filled in.

closes #764