refresh: use exponential retry backoff with base 2 rather than 3
By using the cumulative time since the zone expiration (instead of the last
time interval) we achieve an approximation of exponential retry backoff with
base 2 (rather than with base 3, as it was until now). Now, with every retry attempt,
the interval between refreshes only doubles (with up to 30 second jitter).

This change helps avoid excessive waiting for a refresh when a few early attempts
fail.