pkt: new function to return original wire format QNAME
This commit adds a new function knot_pkt_wire_qname() that returns the
QNAME in the wire format buffer. This is implemented the same as the
original knot_pkt_qname() function before this branch, but now that we
no longer modify the QNAME in the wire format buffer in-place, this
should now be the "original" QNAME as seen off the wire (if the packet
was parsed from a socket buffer).