ctl/backup: change order of direct operations and planned events (direct ops first)
This is necessary for zone reset performed in the end of the backup event handler
to occur after the direct operations (i.e. catalog restore and keysonly restore)
have been finished. Without this change, proper operation wouldn't be guaranteed
in non-blocking mode and restore would never work in blocking mode.