knotd: tune the start-detection/loaded-detection loop
Tune the start-detection loop to be dependent on the system performance and the zonedb size.
The sleep time is set to be double of the time to walk through whole zonedb, with upper and
lower limits.

Sleep time in the loaded-detection loop is constant for one minute, then starts to grow
exponentially. It starts at 1 second or at the start-detection sleep time, whichever is higher,
and stops at an upper limit.